# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The rest api which is developed in node.js for servicing Humally's back-end requirements.  ",
    "version": "1.0.0",
    "title": "Humally Backend Layer",
    "contact": {
      "email": "ertan.caliskan@amatisbilisim.com"
    }
  },
  "host": "humallyservicelayer-env.uumjp5ppd8.eu-west-3.elasticbeanstalk.com",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/getToken": {
      "get": {
        "tags": [
          "Humally"
        ],
        "summary": "Gets token value from user id and secret",
        "description": "",
        "operationId": "getToken",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "description": "user id of user",
            "required": true,
            "type": "string"
          },
          {
            "name": "secret",
            "in": "query",
            "description": "authentication secret of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "400": {
            "description": "Error occured in server-side and client will receive that with type and message"
          },
          "403": {
            "description": "Unauthorized request"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object"
        },
        "success": {
          "type": "boolean"
        },
        "error": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Error"
          }
        }
      },
      "xml": {
        "name": "ApiResponse"
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Error"
      }
    },
...
  }
}